+++ doc/src/history.md
@@ -1,5 +1,9 @@
 # Release history
 
+## Unreleased
+
+- autoWire: Not yet documented.
+
 ## 0.2.0 (2024-10-03)
 
 Initial release, branched from `nixos-flake`